backend/pyproject.toml
[project]
name = "devtrack"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=5.1,<6.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-solo>=2.3,<3.0",
    "django-filter>=24.0,<25.0",
    "psycopg[binary]>=3.2,<4.0",
    "django-page-perms",
    "openai>=2.29.0,<3.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.9,<5.0",
    "factory-boy>=3.3,<4.0",
    "faker>=30.0,<40.0",
]

[tool.uv.sources]
django-page-perms = { path = "../packages/django-page-perms", editable = true }
